Cleaned/commented up and renamed the file for timing throws
Kept a lot of potentially unnecessary code from the MPU6050 example
file.

Main engineering/physics problem with this code: the accelerometer
probably won’t notice it’s in the air (“in free fall”) if the ball is
spinning (it’ll feel a centripetal acceleration toward the center of
the ball, which gets worse the faster the spin and the farther it is
from the ball’s center).

Also, the non-gravity acceleration thresholds for determining when it’s
entering and leaving free fall were chosen somewhat arbitrarily and may
need changing based on the final packaging of the accelerometer.