Added acceleration history array and spike/plunge detection functions for better freefall detection
The acceleration spike/plunge approach to free fall detection is a bit
ad hoc, but it avoids the bug isFreeFalling() and even i2C’s
getIntFreefallStatus() functions have where it’s triggered by simply
tilting and holding the ball a certain way.
We still need to play with historyLength and the thresholdPercentages
put into the accelIsSpiking and accelIsPlunging functions in
lightCycle() for best detection.