New Processing app for user-created light shows (currently only in red)
Offers the user keyframes and sliders and stuff to design an animation
(“light show”) of the color red over time
Worked out most potential bugs.

Still need to integrate with Derin’s Processing app to change the
Arduino color.
Also need to add boxes and code for green and blue.